cloth_size_recommendation/requirements.txt
torch
torchvision
scikit-learn
mediapipe
opencv-python-headless
Pillow
streamlit
numpy